+++ lib/active_record_shards/db_connection_5_1.rb
