+++ lib/active_record_shards/sharding.rb
@@ -123,9 +123,9 @@ end
 
 case "#{ActiveRecord::VERSION::MAJOR}.#{ActiveRecord::VERSION::MINOR}"
 when '5.0'
-  require 'active_record_shards/connection_switcher_5_0'
+  require 'active_record_shards/db_connection_5_0'
 when '5.1', '5.2'
-  require 'active_record_shards/connection_switcher_5_1'
+  require 'active_record_shards/db_connection_5_1'
 else
   raise "ActiveRecordShards is not compatible with #{ActiveRecord::VERSION::STRING}"
 end
